% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meta_info.r
\name{read_meta_info}
\alias{read_meta_info}
\title{Import metadata for a DArT dataset}
\usage{
read_meta_info(dart_data, basedir, species, dataset, version = "db")
}
\arguments{
\item{datafile}{-- name of xls file containing the meta data \link{required}}

\item{nas}{-- missing data character \link{default "-"}}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    -- a dart meta data object, consisting of sample names
}\if{html}{\out{</div>}}
}
\description{
read.meta.data() opens a meta data file associated with a dart dataset
}
\examples{
meta_data <- read_dart_meta(datafile=, nas="-")
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
